Avoid operating on the readr problems attribute directly
In the upcoming readr 2.0.0 release the internals of the problems
attribute have changed. It is no longer a tibble of problems, instead it
is an external pointer to the index that is converted to a tibble by the
`problems()` function.

The public interface to retrieve the problems was always using the
`problems()` function, these code changes in this commit convert the
existing code to use only the public interface.